requirements.txt
amqp==5.3.1
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
cbor2==5.8.0
celery==5.6.2
cffi==2.0.0
channels==4.3.1
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.5
daphne==4.1.2
Django==6.0.2
django-cors-headers==4.9.0
django-debug-toolbar==6.2.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
eventlet==0.40.2
greenlet==3.3.2
gunicorn==22.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httptools==0.7.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.11
imbalanced-learn==0.14.1
Incremental==24.11.0
joblib==1.5.3
kombu==5.6.2
msgpack==1.1.2
numpy==2.4.2
packaging==26.0
pandas==3.0.1
pillow==12.1.1
priority==1.3.0
prompt_toolkit==3.0.52
psycopg[binary]==3.2.13
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.2
PyYAML==6.0.3
redis==6.4.0
scikit-learn==1.8.0
scipy==1.17.0
service-identity==24.2.0
six==1.17.0
sklearn-compat==0.1.5
sqlparse==0.5.5
threadpoolctl==3.6.0
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
uvicorn==0.30.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
whitenoise==6.11.0
zope.interface==8.2
